Here goes nothing, attempt to externalize class, without rebuilding my environment for this minecraft version
